ajout visuel du temps + amélioration fluidité déplacement des monstres et projectiles

+++ code/app/src/main/java/com/androidtowerdefense/model/Coordinate.java
@@ -7,8 +7,8 @@ import java.io.Serializable;
  * Coordonnées
  */
 public class Coordinate implements Serializable {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
     private static int tileWidth;
     private static int tileHeight;
 
@@ -17,7 +17,7 @@ public class Coordinate implements Serializable {
      * @param x int Position X
      * @param y int Position Y
      */
-    public Coordinate(int x , int y){
+    public Coordinate(float x , float y){
         this.x = x;
         this.y = y;
     }
@@ -34,7 +34,7 @@ public class Coordinate implements Serializable {
      * Accède a la position X exacte sur la fenêtre
      * @return int Position X
      */
-    public int getExactX() {
+    public float getExactX() {
         return x * tileWidth + (tileWidth/2);
     }
 
@@ -42,13 +42,13 @@ public class Coordinate implements Serializable {
      * Accède a la position Y exacte sur la fenêtre
      * @return  int Position Y
      */
-    public int getExactY() { return y * tileHeight + (tileHeight/2);}
+    public float getExactY() { return y * tileHeight + (tileHeight/2);}
 
-    public int getX() {return x;}
-    public void setX(int x) {this.x = x;}
+    public float getX() {return x;}
+    public void setX(float x) {this.x = x;}
 
-    public int getY() {return y;}
-    public void setY(int y) {this.y = y;}
+    public float getY() {return y;}
+    public void setY(float y) {this.y = y;}
 
     public boolean equals(Coordinate obj) {
         return this.x == obj.x && this.y == obj.y;
